leftoversIA_API/requirements.txt
Flask==3.0.0
Flask_Cors==4.0.0
python-dotenv==1.0.0
Requests==2.31.0
gunicorn==21.2.0
